split out server code from types package
