+++ x-pack/solutions/search/packages/workchat/wci-server/README.md
@@ -1,0 +1,3 @@
+# @kbn/wci-server
+
+Contains server-side implementation for WorkChat integrations. Uses types from `@kbn/wci-common`.
