+++ x-pack/solutions/search/packages/workchat/wci-server/kibana.jsonc
@@ -1,0 +1,9 @@
+{
+  "type": "shared-server",
+  "id": "@kbn/wci-server",
+  "owner": [
+    "@elastic/search-kibana"
+  ],
+  "group": "search",
+  "visibility": "private"
+}
