+++ x-pack/solutions/search/packages/workchat/wci-server/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@kbn/wci-server",
+  "private": true,
+  "version": "1.0.0",
+  "license": "Elastic License 2.0",
+  "sideEffects": false
+}
