+++ x-pack/solutions/search/packages/workchat/wci-server/src/mcp/sse_client.ts
