+++ x-pack/solutions/search/packages/workchat/wci-server/src/utils/create_external_client.ts
@@ -5,20 +5,15 @@
  * 2.0.
  */
 
-import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { InMemoryTransport } from '@modelcontextprotocol/sdk/inMemory.js';
 import { Client } from '@modelcontextprotocol/sdk/client/index.js';
-import type { IntegrationClient } from '../types';
+import { SSEClientTransport } from '../mcp/sse_client';
+import type { IntegrationClient } from '@kbn/wci-common';
 
-/**
- * Returns a {@link IntegrationClient} that will run the provided server in memory
- * and connect to it.
- */
-export const getClientForInternalServer = async ({
-  server,
+export const getClientForExternalServer = async ({
+  serverUrl,
   clientName = 'unknown',
 }: {
-  server: McpServer;
+  serverUrl: string;
   clientName?: string;
 }): Promise<IntegrationClient> => {
   let _disconnect: () => Promise<void> | undefined;
@@ -31,18 +26,26 @@ export const getClientForInternalServer = async ({
       }
       connected = true;
 
-      const [clientTransport, serverTransport] = InMemoryTransport.createLinkedPair();
-      const client = new Client({
-        name: clientName,
-        version: '1.0.0',
-      });
+      const transport = new SSEClientTransport(new URL(serverUrl));
 
-      await server.connect(clientTransport);
-      await client.connect(serverTransport);
+      const client = new Client(
+        {
+          name: clientName,
+          version: '1.0.0',
+        },
+        {
+          capabilities: {
+            prompts: {},
+            resources: {},
+            tools: {},
+          },
+        }
+      );
+
+      await client.connect(transport);
 
       _disconnect = async () => {
         await client.close();
-        await server.close();
       };
 
       return client;
