+++ x-pack/solutions/search/packages/workchat/wci-server/src/utils/index.ts
