+++ x-pack/solutions/search/plugins/workchat/wci_external_server/server/integration/external_server_integration.ts
@@ -10,8 +10,8 @@ import {
   IntegrationType,
   WorkchatIntegrationDefinition,
   WorkChatIntegration,
-  getClientForExternalServer,
 } from '@kbn/wci-common';
+import { getClientForExternalServer } from '@kbn/wci-server';
 import { WCIExternalServerConfiguration } from '../types';
 
 export const getExternalServerIntegrationDefinition = ({
