+++ x-pack/solutions/search/plugins/workchat/wci_salesforce/server/integration/salesforce_integration.ts
@@ -10,8 +10,8 @@ import {
   IntegrationType,
   WorkchatIntegrationDefinition,
   WorkChatIntegration,
-  getClientForInternalServer,
 } from '@kbn/wci-common';
+import { getClientForInternalServer } from '@kbn/wci-server';
 import { createMcpServer } from './mcp_server';
 import { SalesforceConfiguration } from '../types';
